requirements.txt
discord.py
python-dotenv~=0.20.0
pytz~=2022.1